Fix "NotFoundError" upon restarting game
The game encountered an "Uncaught NotFoundError"
upon restarting the game (two or more times). This
was due to the event listener for gameButton being
added to the resetGame function, which caused a
new event listener to be created every time a new
game was started. The error was shown when event
listeners after the first attempted to remove the
gameButton child, which had already been removed
before by the first event listener.